Reorganize all of the `EventHandler`s and fix some bugs in turn. Make use of the assumption that the items in `Items.kt` are only available in the lobby world. Refactor and simplify a lot of features and make methods more concise. Could possibly be multiple commits but in fear of pushing broken code, this is just one.
