+++ src/main/kotlin/gtlp/groundmc/lobby/LobbyMain.kt
@@ -7,10 +7,7 @@ import gtlp.groundmc.lobby.commands.*
 import gtlp.groundmc.lobby.database.table.Meta
 import gtlp.groundmc.lobby.database.table.Relationships
 import gtlp.groundmc.lobby.database.table.Users
-import gtlp.groundmc.lobby.event.listener.EntityEventListener
-import gtlp.groundmc.lobby.event.listener.InventoryClickEventListener
-import gtlp.groundmc.lobby.event.listener.MiscEventListener
-import gtlp.groundmc.lobby.event.listener.PlayerEventListener
+import gtlp.groundmc.lobby.event.listener.*
 import gtlp.groundmc.lobby.inventory.LobbyInventory
 import gtlp.groundmc.lobby.registry.LobbyCommandRegistry
 import gtlp.groundmc.lobby.task.*
@@ -91,10 +88,7 @@ class LobbyMain : JavaPlugin() {
             Meta.upgradeDatabase()
         }
         logger.finer("Registering events...")
-        Bukkit.getServer().pluginManager.registerEvents(EntityEventListener(), this)
-        Bukkit.getServer().pluginManager.registerEvents(InventoryClickEventListener(), this)
-        Bukkit.getServer().pluginManager.registerEvents(MiscEventListener(), this)
-        Bukkit.getServer().pluginManager.registerEvents(PlayerEventListener(), this)
+        registerListeners()
         registerCommands()
 
         logger.finer("Scheduling tasks...")
@@ -106,6 +100,18 @@ class LobbyMain : JavaPlugin() {
         logger.finer("Setting difficulty of the hub world to peaceful")
         hubLocation.get().world.difficulty = Difficulty.PEACEFUL
         logger.exiting(LobbyMain::class, "onEnable")
+    }
+
+    private fun registerListeners() {
+        arrayOf(ChatInteractionListener,
+                HidePlayerListener,
+                LobbyInteractionListener,
+                LobbyInventoryListener,
+                LobbyInvincibilityListener,
+                PreventWorldInteractionListener,
+                ServerStateListener,
+                SilentChatListener
+        ).forEach { Bukkit.getServer().pluginManager.registerEvents(it, this) }
     }
 
     /**
@@ -292,7 +298,7 @@ class LobbyMain : JavaPlugin() {
         /**
          * Variable to hold the [Location] of the hub/lobby.
          */
-        var hubLocation: Optional<Location> = Optional.empty()
+        lateinit var hubLocation: Optional<Location>
 
         /**
          * A map of tasks to their IDs
