+++ src/main/kotlin/gtlp/groundmc/lobby/event/listener/ChatInteractionListener.kt
@@ -1,0 +1,33 @@
+package gtlp.groundmc.lobby.event.listener
+
+import gtlp.groundmc.lobby.LobbyMain
+import gtlp.groundmc.lobby.util.I18n
+import org.bukkit.event.EventHandler
+import org.bukkit.event.EventPriority
+import org.bukkit.event.Listener
+import org.bukkit.event.player.AsyncPlayerChatEvent
+import org.bukkit.metadata.FixedMetadataValue
+import org.joda.time.Instant
+
+object ChatInteractionListener : Listener {
+
+    /**
+     * Slows down the chat by cancelling fast chat messages.
+     *
+     * @param event the event to handle
+     */
+    @EventHandler(priority = EventPriority.LOWEST)
+    fun slowChat(event: AsyncPlayerChatEvent) {
+        if (LobbyMain.instance.get().config.getBoolean("slowchat.enabled")) {
+            val key = "lastChatMsg"
+            if (event.player.hasMetadata(key) && (Instant.now() - event.player.getMetadata(key).first { it.owningPlugin == LobbyMain.instance.get() }.asLong()) < Instant(LobbyMain.instance.get().config.getLong("slowchat.timeout"))) {
+                event.isCancelled = true
+                event.player.sendMessage(I18n.getString("too_many_messages", event.player.locale))
+                LobbyMain.logger.finest("${event.player} sent messages too quickly!")
+            } else {
+                event.player.setMetadata(key, FixedMetadataValue(LobbyMain.instance.get(), Instant().millis))
+            }
+        }
+    }
+
+}
