+++ src/main/kotlin/gtlp/groundmc/lobby/event/listener/LobbyInvincibilityListener.kt
@@ -1,19 +1,14 @@
 package gtlp.groundmc.lobby.event.listener
 
 import gtlp.groundmc.lobby.LobbyMain
-import gtlp.groundmc.lobby.enums.NBTIdentifier
-import gtlp.groundmc.lobby.util.NBTItemExt
 import org.bukkit.entity.Player
 import org.bukkit.event.EventHandler
 import org.bukkit.event.Listener
-import org.bukkit.event.block.BlockPlaceEvent
+import org.bukkit.event.entity.EntityDamageEvent
+import org.bukkit.event.entity.EntityDamageEvent.DamageCause.*
 import org.bukkit.event.entity.PotionSplashEvent
 
-/**
- * Listener to handle miscellaneous events.
- */
-class MiscEventListener : Listener {
-
+object LobbyInvincibilityListener : Listener {
     /**
      * Removes potion effects from splashed potions in the lobby world.
      *
@@ -28,16 +23,14 @@ class MiscEventListener : Listener {
     }
 
     /**
-     * Prevents players to place blocks with the [NBTIdentifier.PREFIX] on them.
+     * Handles the [EntityDamageEvent] and cancels all damage in the lobby world.
      *
      * @param event the event to handle
      */
     @EventHandler
-    fun cancelBlockPlace(event: BlockPlaceEvent) {
-        val nbtItem = NBTItemExt(event.itemInHand)
-        if (nbtItem.hasKey(NBTIdentifier.PREFIX)!!) {
+    fun preventDamage(event: EntityDamageEvent) {
+        if (event.entity.world == LobbyMain.hubLocation.get().world && event.cause !in arrayOf(SUICIDE, CUSTOM, VOID)) {
             event.isCancelled = true
         }
     }
-
 }
